WIP: game history component (fix scroll target
